Inline CopySmallDist, CopyBigDist and CopySetLength
Inlined the functions CopySmallDist and CopyBigDist. This saves one JP
instruction.

Also save a JP instruction for CopySetLength. Though instead of inlining
this routine I copied it in front of the (generated) DistanceTree code.

Change was suggested by Prodatron on MRC. It saves about 0.2s-0.3s on Z80
(Aleste2 test).